Fix "txpacket() randomly gets stuck on Linux"
It turned out that Python SDK calls termios.tcdrain() to clear remaining data
in the serial port. Since tcdrain() is a blocking operation, it can paralyze
the Python interpreter execution for a few seconds.

On the other hand, C/C++ SDK uses `tcflush()` for the same purpose, so simply
dicards the data not transmitted yet. Fix Python SDK to do the same.

Signed-off-by: Fujimoto Seiji <fujimoto@ceptord.net>